requirements.txt
colorama==0.4.4
crayons==0.3.1
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.20.2
opencv-python==4.5.1.48
piexif==1.1.3
Pillow==6.2.2
stegano==0.9.8
